// src/models/all-proxies-parsed.ts
export interface AllProxiesParsed {
  coolProxy: string[];
  freeProxyList: string[];
  freeProxyLists: string[];
  hideMyName: string[];
  newNetTime: string[];
  premProxy: string[];
  proxy5050: string[];
  proxyList: string[];
  proxyNova: string[];
  proxyServerList24: string[];
  proxyHttp: string[];
  socksList: string[];
  spysOne: string[];
}